==> 01_workbooks/03. normalize_serving_size.ipynb <==
# normalize serving size 

import pandas as pd
import numpy as np


df = pd.read_pickle("../00_data/02_comibined_clean_new_serving_size.pkl")

df

df_norm = df.copy()

col = list(df.columns)

col = [ 'serve_size_new',
        'calories_kcal',
         'total_fat_g',
         'saturated_fat_g',
         'trans_fat_g',
         'cholesterol_mg',
         'sodium_mg',
         'carbohydrates_g',
         'dietary_fiber_g',
         'sugars_g',
         'protein_g']

## Tesing the column-wise calculation

df_norm.dtypes

df_norm[col] = df[col].apply(pd.to_numeric, errors='coerce')
df_norm.dtypes

df_norm["test"] = df_norm["calories_kcal"]/df_norm["serve_size_new"]*100
df_norm

#test has worked out. Now right a loop to calculate the normalized value for all columns

--> Worked out well. 

## Converting the serving size from fl oz to ml

df[col] = df[col].apply(pd.to_numeric, errors='coerce')
df

oz_t_ml = 28.413075

df_ml = df.loc[df["serve_size_unit"].str.contains("fl oz")]
df_g = df.loc[df["serve_size_unit"].str.contains("g")]


df_ml = df_ml.assign(serve_size = df_ml["serve_size_new"].mul(oz_t_ml))
df_ml["serve_size_unit"] = "ml"

df_g = df_g.assign(serve_size = df_g["serve_size_new"].copy())
df_g


df_combined = pd.DataFrame()
df_combined = df_combined.append(df_g).append(df_ml)
df_combined.head()

-------------------------------

## The following section were the other attempts to convert fl oz in ml. All resulted in errors, but don't konw why...

-------------------------------

-------------------------- first attempt

df = df.assign(test = "")
df.head()

df.loc[df["serve_size_unit"].str.contains("fl oz"), "test"] = (df["serve_size_new"] * oz_t_ml)

error: cannot reindex duplicte axis. 

-------------------------- second attempt

temp_df = (df.loc[df["serve_size_unit"].str.contains("fl oz"), "serve_size_new"]) *oz_t_ml 
temp_df

temp_df = pd.DataFrame(temp_df)
temp_df

temporary df containing all rows with fl oz conversion

temp_df = temp_df.rename(columns = {"serve_size_new" : "serve_size_ml"})
temp_df

df = pd.merge(df, temp_df, how="left", left_index=True, right_index=True)
df.merge(right=temp_df, how="left", left_index=True, right_index=True)

-> after merging, the new df has more rows than the original df

df_combined = df_combined.rename(columns = {"serve_size_new" : "serve_size_g_oz"})

df_combined.columns

df_combined = df_combined[['chain', 'category_clean', 'item_name','serve_size',
       'serve_size_unit', 'calories_kcal', 'total_fat_g', 'saturated_fat_g',
       'trans_fat_g', 'cholesterol_mg', 'sodium_mg', 'carbohydrates_g',
       'dietary_fiber_g', 'sugars_g', 'protein_g', 'category',
       'serving_size_g', 'serve_size_g_oz']]

df_combined

df_normalized = df_combined.copy()
df_normalized.dtypes

col_calc = [
     'calories_kcal',
     'total_fat_g',
     'saturated_fat_g',
     'trans_fat_g',
     'cholesterol_mg',
     'sodium_mg',
     'carbohydrates_g',
     'dietary_fiber_g',
     'sugars_g',
     'protein_g']

i = 0
for x in col_calc:
     df_normalized[i] = df_normalized[x] / df_normalized["serve_size"]*100
     i = i+1

df_normalized.dtypes

#rename the new columns

old_col = [0,1,2,3,4,5,6,7,8,9]
new_col = ['calories_kcal_norm',
     'total_fat_norm',
     'saturated_fat_norm',
     'trans_fat_norm',
     'cholesterol_norm',
     'sodium_norm',
     'carbohydrates_norm',
     'dietary_fiber_norm',
     'sugars_norm',
     'protein_norm']
old_new_dict = dict(zip(old_col, new_col))
df_normalized = df_normalized.rename(columns = old_new_dict)

df_normalized

df_normalized = df_normalized.round(2)
# df_normalized = df_normalized["serve_size"].round()
df_normalized.columns

df_normalized = df_normalized[['chain','category_clean', 'item_name', 'serve_size', 'serve_size_unit',
       'calories_kcal', 'total_fat_g', 'saturated_fat_g', 'trans_fat_g',
       'cholesterol_mg', 'sodium_mg', 'carbohydrates_g', 'dietary_fiber_g',
       'sugars_g', 'protein_g','calories_kcal_norm', 'total_fat_norm', 'saturated_fat_norm',
       'trans_fat_norm', 'cholesterol_norm', 'sodium_norm',
       'carbohydrates_norm', 'dietary_fiber_norm', 'sugars_norm',
       'protein_norm']]

df_normalized

## df_normalized is the dataframe combining all restaurants & incl. normalized value

df_normalized.to_pickle("../00_data/03_comibined_clean_normalized.pkl")

df_normalized.to_csv("../00_data/03_comibined_clean_normalized.csv", index=False)
