==> 01_workbooks/04. Category_clean_up.ipynb <==
# There are still some categories to be cleaned

import pandas as pd
import numpy as np

df = pd.read_pickle("../00_data/03_comibined_clean_normalized.pkl")

df.head()

df = df.reset_index(drop=True)
df

df.loc[df["item_name"].str.contains("Honey Mustard Dipping"), "category_clean"] = "sauce"
df.loc[df["item_name"].str.contains("Honey Mustard Dipping")]

df.loc[df["item_name"].str.contains("Ketchup"), "category_clean"] = "sauce"

df.loc[df["item_name"].str.contains("Mayonnaise ")]

df.iloc[97,1] = "sauce"

df.loc[df["item_name"].str.contains("Ranch Dipping"),"category_clean"] = "sauce"

df.loc[df["item_name"].str.contains("Zesty"), "category_clean"] = "sauce"

df.loc[df["item_name"].str.contains("Chicken Fries")]

df.loc[df["item_name"].str.contains("McWrap Chicken"), "category_clean"] = "wrap"

df.loc[df["item_name"].str.contains("Pepperoni")]

df = df.drop(df.index[165])

df.loc[df["item_name"].str.contains("Pepperoni")]

df.loc[269:276, "category_clean"] = "sides"

df.loc[280:282,"category_clean" ] = "salad"

df.loc[283:285,"category_clean" ] = "sauce"

df.loc[(df["chain"] == "burger_king") & (df["item_name"].str.contains("BIG FISH")), "category_clean"] = "burger"

df.loc[241:248, "category_clean"] = "burger" 

df.loc[283, "category_clean"] = "salad" 

df.loc[292:295, "category_clean"] = "sides" 

df.loc[278, "category_clean"] = "sides" 

df.loc[288:291, "category_clean"] = "sides" 

df.loc[341:343, "category_clean"] = "sauce" 

df.loc[338, "category_clean"] = "sauce" 

df.loc[339, "category_clean"] = "sauce" 

df.loc[340, "category_clean"] = "sauce" 

df.loc[110:117, "category_clean"] = "dessert" 

df.loc[110:117,]

df.loc[269, "category_clean"] = "burger"

df.loc[277, "category_clean"] = "sides"

df.loc[(df["chain"] == "burger_king") & (df["item_name"].str.contains("Nuggets"))]

# df.loc[(df["chain"] == "burger_king") & (df["item_name"] == "salad")]

# df.loc[(df["chain"] == "burger_king") & (df["item_name"].str.contains("BIG FISH"))]

df.columns

## Exporting final data frame

df.to_csv("../00_data/04_comibined_final.csv", index=False)

df.to_pickle("../00_data/04_comibined_final.pkl")
