==> 01_workbooks/05 Convert_data_structure.ipynb <==
# Converting the data structure to tidy data format

import pandas as pd
import numpy as np


df = pd.read_pickle("../00_data/04_comibined_final.pkl")

df.reset_index()

## Testing melt with a sample

sample = df.head(20) #use a sample to test the method

sample = sample.iloc[0:10,0:10] #use only the first 10 columns for test

sample.columns

id_vars_list = ["chain", "category_clean", "item_name","serve_size", "serve_size_unit"]

pd.melt(sample, 
            id_vars = id_vars_list).sort_values(by=["chain", "category_clean", "item_name"])

## Use melt to re-structure the absolute values first. 

df.columns

df_per_serving = df[['chain', 'category_clean', 'item_name', 'serve_size', 'serve_size_unit',
       'calories_kcal', 'total_fat_g', 'saturated_fat_g', 'trans_fat_g',
       'cholesterol_mg', 'sodium_mg', 'carbohydrates_g', 'dietary_fiber_g',
       'sugars_g', 'protein_g']]

col_stay = ['chain', 'category_clean', 'item_name', 'serve_size', 'serve_size_unit']

df_per_serving = (pd.melt(df_per_serving, 
                    id_vars = col_stay,
                    value_name = "value_per_serving",
                    var_name = "nutrition_indicator")
                    .sort_values(by=["chain", "category_clean", "item_name"]))

df_per_serving["nutrition_indicator"].value_counts()

## --> as we don't use the normalized value in Tableau, I have dropped the columns here for export! 

df_per_serving.to_pickle("../00_data/05_combined_reshaped_no_norm.pkl")

df_per_serving.to_csv("../00_data/05_combined_reshaped_no_norm.csv", index=False)
