==> 01_workbooks/06. Reshape WHO data.ipynb <==
# Reshape the WHO data to merge into Tableau

import pandas as pd

df = pd.read_csv("../00_data/WHO_recommendation.csv")

df.head()

df = pd.melt(df, id_vars = ["sex", "age"], var_name = "nutrition_indicator", value_name = "rec. daily intake").sort_values(by=["sex", "age"])

df

df.to_csv("../00_data/06_WHO_recommendation_shaped.csv", index=False)
