==> 01_workbooks/08. Deleting "trans_fat" from dataset.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("../00_data/07_combined_clean_no_norm.csv")

df

df = df.loc[df["nutrition_indicator"] != "trans_fat_g"]

df

# deleted the rows including "nutrition_indicator" = "trans_fat" --> because the WHO data does not include the information.

df.to_csv("../00_data/08_combined_data.csv", index=False)

# ATTENTION: 

### Data has been cleaned in excel. 
- For McDonalds salad sauces, the Cholesterol value has been set to 0, as no value available.
- For subway Kidsmenue, the Cholesterol value has been set to 0, as no value available.

---> Version to be used: 08_01_combined_data_excel_cleaned.csv
