==> 01_workbooks/09. Update subway category .ipynb <==
import pandas as pd

df = pd.read_csv("../00_data/08_01_combined_data_excel_cleaned.csv")

df.loc[(df["category_clean"] == "sandwich"), "category_clean"] = "burger/sandwich"

df.loc[(df["category_clean"] == "burger"), "category_clean"] = "burger/sandwich"

df["category_clean"].value_counts()

df.to_csv("../00_data/09_combined_final.csv", index=False)
